Exercice 4 terminé Adam
